Created controller.auth
Created controller.auth and reorganize _login and _updatePassword
